This file was used to install OpenVPN
It is fairly easy to use.  It has interactive prompts.  It installs and configures the server side of OpenVPN, and it produces the single configuration file needed for a client server to use the OpenVPN server.  The client configuration would otherwise be difficult.  The client server must have the openvpn binary installed.  To start the service the OpenVPN server's service must be running.  The client server must run the openvpn command and use a flag to point to the configuration file.